Remove monkey patch and replace with nested class.
